requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
psycopg2-binary==2.9.9
pycodestyle==2.12.1
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
